+++ test/conformance/gateway_conformance_test.go
@@ -24,12 +24,27 @@ import (
 var skippedTestsForTraditionalRoutes = []string{
 	// core conformance
 	tests.HTTPRouteHeaderMatching.ShortName,
+	// There is an issue with KIC when processing this scenario.
+	// TODO: https://github.com/Kong/kubernetes-ingress-controller/issues/6136
+	tests.GRPCRouteListenerHostnameMatching.ShortName,
+	// tests.GRPCRouteHeaderMatching.ShortName and tests.GRPCExactMethodMatching.ShortName may
+	// have some conflicts, skipping either one will still pass normally.
+	// TODO: https://github.com/Kong/kubernetes-ingress-controller/issues/6144
+	tests.GRPCExactMethodMatching.ShortName,
+}
+
+var skippedTestsForExpressionRoutes = []string{
+	// When processing this scenario, the Kong's expressions router requires `priority`
+	// to be specified for routes.
+	// We cannot provide that for routes that are part of the conformance suite.
+	tests.GRPCRouteListenerHostnameMatching.ShortName,
 }
 
 var traditionalRoutesSupportedFeatures = []features.SupportedFeature{
 	// core features
 	features.SupportGateway,
 	features.SupportHTTPRoute,
+	features.SupportGRPCRoute,
 	// extended features
 	features.SupportHTTPRouteResponseHeaderModification,
 	features.SupportHTTPRoutePathRewrite,
@@ -43,6 +58,7 @@ var expressionRoutesSupportedFeatures = []features.SupportedFeature{
 	// core features
 	features.SupportGateway,
 	features.SupportHTTPRoute,
+	features.SupportGRPCRoute,
 	// extended features
 	features.SupportHTTPRouteQueryParamMatching,
 	features.SupportHTTPRouteMethodMatching,
@@ -70,6 +86,7 @@ func TestGatewayConformance(t *testing.T) {
 		supportedFeatures = traditionalRoutesSupportedFeatures
 		mode = string(dpconf.RouterFlavorTraditionalCompatible)
 	case dpconf.RouterFlavorExpressions:
+		skippedTests = skippedTestsForExpressionRoutes
 		supportedFeatures = expressionRoutesSupportedFeatures
 		mode = string(dpconf.RouterFlavorExpressions)
 	default:
@@ -86,6 +103,7 @@ func TestGatewayConformance(t *testing.T) {
 	opts.SkipTests = skippedTests
 	opts.ConformanceProfiles = sets.New(
 		suite.GatewayHTTPConformanceProfileName,
+		suite.GatewayGRPCConformanceProfileName,
 	)
 	opts.Implementation = conformancev1.Implementation{
 		Organization: metadata.Organization,